% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{a name; default is 'world'}
}
\value{
character a string; default is 'Hello, world!'
}
\description{
A one liner.
}
\details{
Here more details
}
\examples{
# Example usage of 'hello'

hello() 
}
\author{
author
}

