% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine_sweeper.R
\name{input.data}
\alias{input.data}
\title{Play the Mine Sweeper game in R
Just type in start() and the game will start
The controls should be familiar to you: Click the left mouse button to dig in
an area, and right button to mark or unmark the area with flags.}
\usage{
input.data()
}
\arguments{
\item{width}{number of grids in horizontal axis}

\item{height}{number of grids in vertical axis}

\item{mines}{number of mines}

\item{cheat}{logical. If \code{TRUE} a matrix indicating the mines will be
printed}
}
\description{
Play the Mine Sweeper game in R
Just type in start() and the game will start
The controls should be familiar to you: Click the left mouse button to dig in
an area, and right button to mark or unmark the area with flags.
}
\note{
Linux/Mac users have to use \code{X11(type = 'Xlib')} or the Cairo
  graphics device \code{Cairo()} in the package \pkg{cairoDevice}.
}
\examples{
should use input to set the size of the map in mine_sweeper
}
\references{
\url{http://en.wikipedia.org/wiki/Minesweeper_(computer_game)}
}
\author{
Yixuan Qiu \email{yixuan.qiu@cos.name}
}
